// types/subgraphs/blocklytics.d.ts
export type Block = {
    id?: string,
    number?: string,
    timestamp?: string,
    parentHash?: string,
    author?: string,
    difficulty?: string,
    totalDifficulty?: string,
    gasUsed?: string,
    gasLimit?: string,
    receiptsRoot?: string,
    transactionsRoot?: string,
    stateRoot?: string,
    size?: string,
    unclesHash?: string
}